refactor display_and_save_batch to utils
